requirements.txt
PyYAML==5.3.1
librosa==0.8.1
numpy==1.20.0
matplotlib==3.4.2
torch==1.9.0
torchvision==0.10.0
torchaudio==0.9.0
spotipy==2.18.0
pydub==0.25.1
Click==8.0.1
py4j==0.10.9.2
xmltodict==0.12.0
tqdm==4.62.2
pytorch-lightning==1.4.0
pandas==1.3.2